api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
authlib>=1.3.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
celery[redis]>=5.4.0
redis>=5.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
google-generativeai>=0.8.0
pgvector>=0.3.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
